#11 Refactored and fixed naming
