+++ tests/Unit/Bus/Middleware/MiddlewareChainTestCase.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace AwdStudio\Tests\Unit\Bus\Middleware;
 
-use AwdStudio\Bus\Handlers;
+use AwdStudio\Bus\HandlerLocator;
 use AwdStudio\Bus\Middleware\MiddlewareChain;
 use AwdStudio\Tests\BusTestCase;
 
@@ -13,14 +13,14 @@ abstract class MiddlewareChainTestCase extends BusTestCase
     /** @var \AwdStudio\Bus\Middleware\MiddlewareChain */
     protected $instance;
 
-    /** @var \AwdStudio\Bus\Handlers|\Prophecy\Prophecy\ObjectProphecy */
+    /** @var \AwdStudio\Bus\HandlerLocator|\Prophecy\Prophecy\ObjectProphecy */
     protected $handlersProphecy;
 
     protected function setUp(): void
     {
         parent::setUp();
 
-        $this->handlersProphecy = $this->prophesize(Handlers::class);
+        $this->handlersProphecy = $this->prophesize(HandlerLocator::class);
 
         $this->instance = new MiddlewareChain($this->handlersProphecy->reveal());
     }
