+++ tests/Unit/Bus/MiddlewareBusTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace AwdStudio\Tests\Unit\Bus;
 
 use AwdStudio\Bus\MiddlewareBus;
-use AwdStudio\Bus\Handlers;
+use AwdStudio\Bus\HandlerLocator;
 use AwdStudio\Bus\Middleware;
 use AwdStudio\Tests\BusTestCase;
 use Prophecy\Argument;
@@ -18,7 +18,7 @@ final class MiddlewareBusTest extends BusTestCase
     /** @var \AwdStudio\Bus\MiddlewareBus */
     private $instance;
 
-    /** @var \AwdStudio\Bus\Handlers|\Prophecy\Prophecy\ObjectProphecy */
+    /** @var \AwdStudio\Bus\HandlerLocator|\Prophecy\Prophecy\ObjectProphecy */
     private $handlersProphecy;
 
     /** @var \AwdStudio\Bus\Middleware|\Prophecy\Prophecy\ObjectProphecy */
@@ -28,7 +28,7 @@ final class MiddlewareBusTest extends BusTestCase
     {
         parent::setUp();
 
-        $this->handlersProphecy = $this->prophesize(Handlers::class);
+        $this->handlersProphecy = $this->prophesize(HandlerLocator::class);
         $this->middlewareProphecy = $this->prophesize(Middleware::class);
 
         $this->instance = new class(
