+++ tests/Unit/Command/CommandBusTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace AwdStudio\Tests\Unit\Command;
 
 use AwdStudio\Bus\Exception\NoHandlerDefined;
-use AwdStudio\Bus\Handlers;
+use AwdStudio\Bus\HandlerLocator;
 use AwdStudio\Bus\Middleware;
 use AwdStudio\Command\CommandBus;
 use AwdStudio\Command\ICommandBus;
@@ -20,7 +20,7 @@ final class CommandBusTest extends BusTestCase
     /** @var \AwdStudio\Command\CommandBus */
     private $instance;
 
-    /** @var \AwdStudio\Bus\Handlers|\Prophecy\Prophecy\ObjectProphecy */
+    /** @var \AwdStudio\Bus\HandlerLocator|\Prophecy\Prophecy\ObjectProphecy */
     private $handlersProphecy;
 
     /** @var \AwdStudio\Bus\Middleware|\Prophecy\Prophecy\ObjectProphecy */
@@ -30,7 +30,7 @@ final class CommandBusTest extends BusTestCase
     {
         parent::setUp();
 
-        $this->handlersProphecy = $this->prophesize(Handlers::class);
+        $this->handlersProphecy = $this->prophesize(HandlerLocator::class);
         $this->middlewareProphecy = $this->prophesize(Middleware::class);
 
         $this->instance = new CommandBus($this->handlersProphecy->reveal(), $this->middlewareProphecy->reveal());
